+++ examples.go
@@ -2,10 +2,11 @@
 
 package main
 
+
 import "fmt"
 
 const (
-	examples = `
+    examples = `
 
 Examples
 
@@ -254,5 +255,6 @@ The default character set is utf-8
 
 // Print Examples ...
 func PrintExamples() {
-	fmt.Println(examples)
+    fmt.Println(examples)
 }
+